% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail_dataset.R
\docType{data}
\name{avail_dataset}
\alias{avail_dataset}
\title{Link to dataset available}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 389 rows and 3 columns.
}
\source{
\url{https://gso.gov.vn}
}
\usage{
avail_dataset
}
\description{
Link to dataset available
}
\keyword{datasets}
