% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gso_avail.R
\name{gso_avail}
\alias{gso_avail}
\title{Retrieve available data set}
\usage{
gso_avail()
}
\value{
A tibble contain all data set tittle from gso.gov.vn
}
\description{
Get available data on https://www.gso.gov.vn/so-lieu-thong-ke/
}
