% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gso_read.R
\name{gso_read}
\alias{gso_read}
\title{Load data set from URL}
\usage{
gso_read(url)
}
\arguments{
\item{url}{link to the dataset from gso_avail()}
}
\value{
A list that contain title and dataset
}
\description{
This function return a dataset retrieve from gso.gov.vn from user provided URL
}
